updated flower to have double properties. instead of int properties.

+++ Flower.cpp
@@ -5,7 +5,7 @@ void Flower::setLabel(std::string newLabel)
     m_label = std::move(newLabel);
 }
 
-Flower::Flower(std::array<int, 4> properties, std::string label)
+Flower::Flower(std::array<double, 4> properties, std::string label)
 {
     m_label = std::move(label);
     m_properties = properties;
